Create historical_stock_revenue_data_analysis.ipynb
This Jupyter notebook, named historical_stock_revenue_data_analysis.ipynb, serves as the core of the Historical_Stock_Revenue_Analysis_Dashboard project. It conducts a thorough analysis of historical stock and revenue data, uncovering trends, patterns, and relationships within the financial data. The notebook also includes code for building an interactive dashboard using tools such as yfinance to fetch financial data, pandas, to manipulate and analyze data and Plotly to create interactive visualizations. Immerse yourself in the realm of financial analytics and visualization with this comprehensive exploration of historical stock and revenue data.